# swagger.json
{"openapi":"3.0.0","info":{"title":"Cosmetic Api","description":"The Cosmetic API","version":"1.0","contact":{}},"tags":[{"name":"Cosmetic","description":""}],"servers":[],"components":{"schemas":{"CreateUserInput":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"facebookId":{"type":"string"},"userType":{"type":"string","default":"customer"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"birthday":{"type":"string"},"avatar":{"type":"string"}},"required":["username","password","facebookId","userType","name","email","phone","address","birthday","avatar"]},"UpdateUserInput":{"type":"object","properties":{"id":{"type":"string"},"password":{"type":"string"},"facebookId":{"type":"string"},"userType":{"type":"string","default":"customer"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"birthday":{"type":"string"},"avatar":{"type":"string"},"isActive":{"type":"boolean"}},"required":["id","password","facebookId","userType","name","email","phone","address","birthday","avatar","isActive"]},"LoginInput":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"CreateProductInput":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"size":{"type":"number"},"color":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"point":{"type":"number"},"detail":{"type":"string"},"promotion":{"type":"number"},"producerId":{"type":"string"},"categoryIds":{"type":"array","items":{"type":"string"}}},"required":["code","name","size","price","quantity","producerId","categoryIds"]},"UpdateProductInput":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"size":{"type":"number"},"color":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"quantity":{"type":"number"},"point":{"type":"number"},"detail":{"type":"string"},"promotion":{"type":"number"},"producerId":{"type":"string"},"categoryIds":{"type":"array","items":{"type":"string"}},"id":{"type":"string"}},"required":["code","name","size","price","quantity","producerId","categoryIds","id"]},"CreateCategory":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"isShowOnSite":{"type":"boolean"},"parentId":{"type":"string"}},"required":["name"]},"UpdateCategoryInput":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"isShowOnSite":{"type":"boolean"},"parentId":{"type":"string"},"id":{"type":"string"}},"required":["name","id"]}}},"paths":{"/user/{id}":{"get":{"operationId":"UserController_find","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[{"name":"offset","required":false,"in":"query","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"limit user","schema":{"default":10,"type":"number"}},{"name":"searchText","required":false,"in":"query","nullable":true,"schema":{"default":"","type":"string"}},{"name":"userType","required":false,"in":"query","nullable":true,"schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserInput"}}}},"responses":{"201":{"description":""}}},"put":{"operationId":"UserController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserInput"}}}},"responses":{"200":{"description":""}}}},"/auth":{"post":{"operationId":"AuthControler_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginInput"}}}},"responses":{"201":{"description":""}}}},"/product":{"get":{"operationId":"ProductController_get","parameters":[{"name":"offset","required":false,"in":"query","schema":{"default":0,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"searchText","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","nullable":true,"schema":{"type":"string"}},{"name":"producer","required":false,"in":"query","nullable":true,"schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductInput"}}}},"responses":{"201":{"description":""}}},"put":{"operationId":"ProductController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductInput"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductController_delete","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/product/create-many":{"post":{"operationId":"ProductController_createMany","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateProductInput"}}}}},"responses":{"201":{"description":""}}}},"/category":{"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategory"}}}},"responses":{"201":{"description":""}}},"put":{"operationId":"CategoryController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryInput"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"CategoryController_delete","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}}}